+++ lib/src/parsers/phone_parser.dart
@@ -4,7 +4,9 @@ import 'package:phone_numbers_parser/src/parsers/_dial_code_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_international_prefix_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_national_prefix_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/light_phone_parser.dart';
+import 'package:phone_numbers_parser/src/parsers/parser.dart';
 import 'package:phone_numbers_parser/src/utils/_metadata_finder.dart';
+import 'package:phone_numbers_parser/src/utils/_metadata_matcher.dart';
 
 import '_iso_code_parser.dart';
 import '_text_parser.dart';
@@ -24,7 +26,7 @@ import '_validator.dart';
 /// size when imported
 ///
 /// It also furnishes more utilities that cannot be achieved with the light parser
-class PhoneParser extends LightPhoneParser {
+class PhoneParser extends Parser {
   /// parses a [phoneNumber] given an [isoCode]
   ///
   /// {@macro phoneNumber}
@@ -38,8 +40,9 @@ class PhoneParser extends LightPhoneParser {
     phoneNumber = InternationalPrefixParser.removeInternationalPrefix(
         phoneNumber, metadata);
     phoneNumber = DialCodeParser.removeDialCode(phoneNumber, metadata.dialCode);
-    final nsn = NationalPrefixParser.transformLocalNsnToInternational(
-        phoneNumber, metadata);
+    final nsn =
+        NationalPrefixParser.transformLocalNsnToInternationalUsingPatterns(
+            phoneNumber, metadata);
     return PhoneNumber(isoCode: metadata.isoCode, nsn: nsn);
   }
 
@@ -51,6 +54,7 @@ class PhoneParser extends LightPhoneParser {
   /// {@macro phoneNumber}
   ///
   /// throws a PhoneNumberException if the dial code is invalid
+  @override
   PhoneNumber parseWithDialCode(String dialCode, String phoneNumber) {
     dialCode = DialCodeParser.normalizeDialCode(dialCode);
     phoneNumber = TextParser.normalize(phoneNumber);
@@ -58,10 +62,13 @@ class PhoneParser extends LightPhoneParser {
         InternationalPrefixParser.removeInternationalPrefix(phoneNumber);
     phoneNumber = DialCodeParser.removeDialCode(phoneNumber, dialCode);
     final metadatas = MetadataFinder.getMetadatasForDialCode(dialCode);
-    final metadata = DialCodeParser.selectMetadataMatchForDialCode(
-        dialCode, phoneNumber, metadatas);
-    final nsn = NationalPrefixParser.transformLocalNsnToInternational(
-        phoneNumber, metadata);
+    final metadata =
+        MetadataMatcher.getMatchUsingPatterns(phoneNumber, metadatas);
+    final nsn =
+        NationalPrefixParser.transformLocalNsnToInternationalUsingPatterns(
+      phoneNumber,
+      metadata,
+    );
     return PhoneNumber(isoCode: metadata.isoCode, nsn: nsn);
   }
 
@@ -73,6 +80,7 @@ class PhoneParser extends LightPhoneParser {
   /// This assumes the phone number starts with the dial code
   ///
   /// throws a PhoneNumberException if the dial code is invalid
+  @override
   PhoneNumber parseRaw(String phoneNumber) {
     phoneNumber = TextParser.normalize(phoneNumber);
     phoneNumber =
@@ -87,10 +95,5 @@ class PhoneParser extends LightPhoneParser {
   @override
   bool validate(PhoneNumber phoneNumber, [PhoneNumberType? type]) {
     return Validator.validateWithPattern(phoneNumber, type);
-  }
-
-  @override
-  PhoneNumber copyWithIsoCode(PhoneNumber phoneNumber, String isoCode) {
-    return parseWithIsoCode(isoCode, phoneNumber.nsn);
   }
 }