+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 1.1.0
+## 2.0.0
 - Resolve issue #1 where it was possible to significantly increase memory foothprint by importing both parsers. Thanks @xvrh 
+- Deplaced validate method from PhoneNumber to PhoneParser and LightPhoneParser to 
+  make it more obvious that one is 
+- Light parser can now parseWithDialCode and parseRaw, although less accurate than the full parser
 
 ## 1.0.5
 - Minor doc change